DCS-835: Validation module moved into a shared folder inside routes folder and re-referenced modules that were dependant on it
